+++ lib/event.dart
@@ -1,3 +1,3 @@
-export 'event/element.dart';
+export 'event/message.dart';
 export 'event/queue.dart';
 export 'event/watcher.dart';